fixed bugs at konoha.class (based on wakamori bugs report)
