+++ package/konoha.class/CMakeLists.txt
@@ -1,4 +1,3 @@
 cmake_minimum_required(VERSION 2.6)
 set(PACKAGE_SOURCE_CODE class_glue.c)
-set(PACKAGE_SCRIPT_CODE class_glue.k)
 add_konoha_package(konoha.class)
